+++ src/settings/OrderTemplates/OrderTemplateView/OrderTemplateView.test.js
@@ -1,7 +1,14 @@
-import { MemoryRouter } from 'react-router';
+import {
+  QueryClient,
+  QueryClientProvider,
+} from 'react-query';
+import { MemoryRouter } from 'react-router-dom';
 
+import {
+  render,
+  screen,
+} from '@folio/jest-config-stripes/testing-library/react';
 import user from '@folio/jest-config-stripes/testing-library/user-event';
-import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
 
 import {
   HasCommand,
@@ -65,12 +72,22 @@ const defaultProps = {
   stripes: { hasPerm: jest.fn().mockReturnValue(true) },
 };
 
+const queryClient = new QueryClient();
+
+const wrapper = ({ children }) => (
+  <MemoryRouter>
+    <QueryClientProvider client={queryClient}>
+      {children}
+    </QueryClientProvider>
+  </MemoryRouter>
+);
+
 const renderOrderTemplateView = (props = {}) => render(
   <OrderTemplateView
     {...defaultProps}
     {...props}
   />,
-  { wrapper: MemoryRouter },
+  { wrapper },
 );
 
 describe('OrderTemplateView', () => {