+++ src/settings/OrderTemplates/OrderTemplatesEditor/DonorInformationForm/DonorInformationForm.js
@@ -1,0 +1,55 @@
+import get from 'lodash/get';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import { Loading } from '@folio/stripes/components';
+import {
+  Donors,
+  fundDistributionShape,
+  useFunds,
+  VisibilityControl,
+} from '@folio/stripes-acq-components';
+
+import { POL_FORM_FIELDS } from '../../../../components/POLine/const';
+import { useManageDonorOrganizationIds } from '../../../../components/POLine/hooks';
+
+const DonorInformationForm = ({ formValues, fundDistribution }) => {
+  const initialDonorOrganizationIds = get(formValues, POL_FORM_FIELDS.donorOrganizationIds, []);
+
+  const {
+    funds,
+    isLoading: isFundsLoading,
+  } = useFunds();
+
+  const {
+    donorOrganizationIds,
+    onDonorRemove,
+    setDonorIds,
+  } = useManageDonorOrganizationIds({
+    funds,
+    fundDistribution,
+    initialDonorOrganizationIds,
+  });
+
+  if (isFundsLoading) {
+    return <Loading />;
+  }
+
+  return (
+    <VisibilityControl name="hiddenFields.donorsInformation">
+      <Donors
+        donorOrganizationIds={donorOrganizationIds}
+        name="donorOrganizationIds"
+        onChange={setDonorIds}
+        onRemove={onDonorRemove}
+      />
+    </VisibilityControl>
+  );
+};
+
+DonorInformationForm.propTypes = {
+  formValues: PropTypes.object.isRequired,
+  fundDistribution: fundDistributionShape,
+};
+
+export default DonorInformationForm;